restructure form
