+++ src/views/manage/components/FormInput.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export const FormInput = ({ name, Component, ...otherProps }) => {
+  return <Component name={name} {...otherProps} />;
+};
+
+export default FormInput;
